Server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "access-needs-agent"
version = "0.1.0"
description = "Multi-channel audio ASR pipeline for accessibility agent"
requires-python = ">=3.11"
dependencies = [
    "deepgram-sdk>=3.0,<4",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "soundfile>=0.12",
    "srt>=3.5",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
pipeline = "pipeline.main:cli"
watcher = "watcher.process_pending:main"
api = "api_server.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
